RouterModule now with providers. Default router-outlet example.
